+++ Storage/SessionStorage.php
@@ -5,11 +5,12 @@ declare(strict_types=1);
 namespace Andreo\OAuthClientBundle\Storage;
 
 use Andreo\OAuthClientBundle\Client\HttpContext;
-use Andreo\OAuthClientBundle\Exception\ExpiredAccessTokenException;
+use Andreo\OAuthClientBundle\Exception\ObjectExpiredException;
 use Andreo\OAuthClientBundle\Exception\StorableNotExistException;
 use Andreo\OAuthClientBundle\Storage\Encoder\EncoderInterface;
 use Andreo\OAuthClientBundle\Storage\Serializer\SerializerInterface;
 use DateTimeImmutable;
+use LogicException;
 
 final class SessionStorage implements StorageInterface
 {
@@ -41,10 +42,16 @@ final class SessionStorage implements StorageInterface
         $string = $request->getSession()->get($key);
 
         $storable = $this->serializer->deserialize($this->encoder::decode($string));
+        if (!$storable->mayBeExpired()) {
+            return $storable;
+        }
+        if (!$storable instanceof ThisIsExpiringInterface) {
+            throw new LogicException('Expiring object must implements '.ThisIsExpiringInterface::class);
+        }
         if ($storable->getExpiredAt() <= new DateTimeImmutable()) {
             $this->delete($httpContext, $key);
 
-            throw new ExpiredAccessTokenException($storable);
+            throw new ObjectExpiredException($storable);
         }
 
         return $storable;
